Restrcure code and add an about page
* Use a clearer directory structure - e.g. external code in
  a different directory
* Split core.php into multiple files each contaning similar functions
* Minor changes to some functions' structure
* Add an 'about' page for when there is no valid item code, with an
  example link - a nicer entry point if no query params in the url
