+++ css/styles.css
@@ -1,0 +1,133 @@
+* {
+    box-sizing: border-box;
+	font-family: "Verdana", sans-serif;
+}
+
+body {
+    margin: 0;
+}
+
+/* Style the header */
+.header {
+	color: #ffffff;
+    background-color: #001372;
+    padding: 50px;
+    text-align: center;
+}
+
+/* Responsive layout: Will display a grid with one to four columns based on device width.
+   Html should be laid out as:
+	<div class='flex-grid'>
+		<div class='flex-cell'>
+			<div class='flex-content-wrapper'>
+				<div class='flex-content'>Content1 content1 content1...</div>
+			</div>
+		</div>
+		<div class='flex-cell'>
+			<div class='flex-content-wrapper'>
+				<div class='flex-content'>Content2 content2 content2...</div>
+			</div>
+		</div>
+		... repeat as required ...
+	</div>
+ */
+.flex-grid {
+	background: #ccf;
+	display: flex;
+	flex-wrap: wrap;
+	padding: 5px;
+}
+.flex-cell {
+    display: flex;
+	width: 25%;
+}
+@media (max-width: 1000px) {
+    .flex-cell {
+		width: 33.33%;
+    }
+}
+@media (max-width: 750px) {
+    .flex-cell {
+		width: 50%;
+    }
+}
+@media (max-width: 600px) {
+    .flex-cell {
+		width: 100%;
+    }
+}
+
+/* Spacing around boxes in the grid, expands to fill available height */
+.flex-content-wrapper {
+	padding: 10px;
+	display: flex;
+    width: 100%;
+    align-content: stretch;
+}
+
+.flex-content {
+    background-color: #eee;
+    border:2px solid #36c;
+	padding: 5px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    border-radius: 10px;
+	width:100%;
+}
+
+a.flex-content:hover, a.flex-content:active {
+    background-color: #fff;
+}
+
+/* Style the footer */
+.footer {
+    background-color: #f1f1f1;
+    padding: 10px;
+    text-align: center;
+}
+
+/* Text styles */
+.main-label {
+    background-color: #2a4b8d;
+	color: #fff;
+	font-size: 140%;
+	font-weight: bold;
+	padding: 2px 10px;
+}
+.main-desc {
+    background-color: #72777d;
+	color: #fff;
+	font-size: 120%;
+	padding: 2px 20px;
+}
+.label {
+	font-size: 110%;
+	font-weight: bold;
+	color: #36c;
+}
+.desc {
+	font-size: 90%;
+	color: #555;
+	line-height: 1.1em;
+	margin-top: 0.2em;
+}
+.small {
+	font-size: 80%;
+}
+code {
+	font-family: monospace;
+}
+
+/* Language selection, Edit on wikidata */
+.main-extra {
+	float:right;
+	font-size:85%;
+	height:30px;
+}
+
+/* Logo images */
+.logo {
+	width: 50px;
+	float:left;
+}
