+++ img/CREDITS
@@ -1,0 +1,37 @@
+Reasonator icon <reasonator.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Reasonator_logo_proposal.png>
+ Author(s): CristianCantoro
+ License: CC-BY-SA 3.0 Unported license <https://creativecommons.org/licenses/by-sa/3.0/deed.en>
+
+Language selector icon <language_selection.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:WMF-Agora-Language_selection-crop_3492D1.svg> 
+ Author(s): Pau Giner
+ License: CC0 Universal Public Domain Dedication <https://creativecommons.org/publicdomain/zero/1.0/deed.en>
+
+Wikibooks icon <wikibooks.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Notification-icon-Wikibooks-logo.svg>
+ Author(s): User:Jdforrester (WMF), User:Ramac, User:Bastique
+ License: CC-BY-SA 3.0 Unported license <https://creativecommons.org/licenses/by-sa/3.0/deed.en>
+
+Wikimedia Commons icon <commons.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Notification-icon-Commons-logo.svg>
+ Author(s): User:Jdforrester (WMF), User:3247, User:Reidab
+ License: Public domain
+
+Wikipedia icon <wikipedia.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Notification-icon-Wikipedia-logo.svg>
+ Author(s): User:Jdforrester (WMF), Jonathan Hoefler
+ License: Public domain
+
+Wikisource icon <wikisource.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Notification-icon-Wikisource-logo.svg>
+ Author(s): User:Jdforrester (WMF), User:Rei-artur, Nicholas Moreau
+ License: CC-BY-SA 3.0 Unported license <https://creativecommons.org/licenses/by-sa/3.0/deed.en>
+
+Wikivoyage icon <wikivoyage.png>:
+ Source: <https://commons.wikimedia.org/wiki/File:Notification-icon-Wikivoyage-logo.svg>
+ Author(s): User:Jdforrester (WMF), User:AleXXw, User:‎Danapit
+ License: CC-BY-SA 3.0 Unported license <https://creativecommons.org/licenses/by-sa/3.0/deed.en>
+ 
+Note:
+ Wikimedia site icons ™ Wikimedia Foundation, Inc. These images are used in this project under the terms of the Trademark policy, section 3.6 <https://wikimediafoundation.org/wiki/Trademark_policy>
