+++ inc/external/SOURCES
@@ -1,0 +1,1 @@
+getDefaultLanguage: <https://www.dyeager.org/2008/10/getting-browser-default-language-php.html>
