+++ inc/helpers.php
@@ -1,0 +1,69 @@
+ <?php
+/* ---------- Helper functions ------------------------------------------------------------------ */
+// Returns the base language (i.e. `en` for `en-gb`), except for language variants which have a
+// Wikipedia.
+function getBaseLanguage($code) {
+	if ( !strpos($code, '-') ) {
+		return $code;
+	}
+	
+	switch ($code) {
+		// Languages with dashes which have Wikipedias can be returned as-is:
+		case 'zh-yue':
+		case 'roa-rup':
+		case 'map-bms':
+		case 'nds-nl':
+		case 'bat-smg':
+		case 'roa-tara':
+		case 'fiu-vro':
+		case 'zh-min-nan':
+		case 'zh-classical':
+		case 'cbk-zam':
+		case 'be-x-old':
+			return $code;
+			break;
+		default:
+			// Strip everything after and including the first dash
+			return explode("-", $code)[0];
+	}
+}
+
+// Get data from deeply nested arrays, or a default value if any of the nested keys aren't set;
+// i.e. `getDeepData($arr, [key1, key2, key3])` is equivalent to `$arr[key1][key2[key3]`
+// if all the keys are set.
+function getDeepData($arr, $keys, $default="") {
+	$data = $arr;
+	foreach ( $keys as $key ) {
+		if ( !isset($data[$key]) ) {
+			return $default;
+		}
+		$data = $data[$key];
+	}
+	return $data;
+}
+
+function getSiteType($site_code) {
+	if ( $site_code === 'commonswiki' ) {
+		return 'commons';
+	}
+	if (strpos($site_code, 'wikisource') !== false) {
+		return 'wikisource';
+	}
+	if (strpos($site_code, 'wikivoyage') !== false) {
+		return 'wikivoyage';
+	}
+	return 'wikipedia';
+}
+
+function extractPageTitle($page) {
+	return $page["title"];
+}
+
+function joinWithPipes($v1, $v2) {
+	if ( $v1 === "" ) {
+		return $v2;
+	}
+	return "{$v1}|{$v2}";
+}
+
+?>
